Let users attach Busy devices as well.
At first I was thinking that Busy meant that it shouldn't be attached to
a VirtualBox VM, but I've found plenty of instances where Busy just means
that the host OS has enumerated it and is ready to be used. This does not
actually prevent the guest OS from using it at all.